split shadow module tests into their own file
