+++ corehq/apps/app_manager/tests/test_suite.py
@@ -644,12 +644,3 @@ class SuiteTest(SimpleTestCase, TestXmlMixin, SuiteMixin):
             'reports.ip1bjs8xtaejnhfrbzj2r6v1fi6hia4i=CommBugz',
             app.create_app_strings('default'),
         )
-
-    def test_shadow_module(self):
-        self._test_generic_suite('shadow_module')
-
-    def test_shadow_module_forms_only(self):
-        self._test_generic_suite('shadow_module_forms_only')
-
-    def test_shadow_module_cases(self):
-        self._test_generic_suite('shadow_module_cases')
